Add files via upload
: A link state packet (LSP) can carry a large
amount of information. For this project, we consider that it carries a minimum
amount of data: the node identity and the cost of links. Both of these details are
needed to make the
topology. LSPs are generated either on a periodic basis or when there is a change in the
topology of the domain.Also called shortest path first (SPF) forwarding.Named after
Dijkstra’s algorithm  which it uses to compute routes.All routers have tables
which contain a representation of the entire network topology.In the form of lists of
routers and information about each router’s neighbours